// src/slices/types.ts
export type SerializedBigNumber = string | null;

interface SerializedFarmUserData {
  allowance: string;
  tokenBalance: string;
  stakedBalance: string;
  earnings: string;
}

export interface DeserializedFarmUserData {
  allowance: BigNumber;
  tokenBalance: BigNumber;
  stakedBalance: BigNumber;
  earnings: BigNumber;
}

export interface SerializedFarm extends SerializedFarmConfig {
  tokenPriceBusd?: string;
  quoteTokenPriceBusd?: string;
  tokenAmountTotal?: SerializedBigNumber;
  quoteTokenAmountTotal?: SerializedBigNumber;
  lpTotalInQuoteToken?: SerializedBigNumber;
  lpTotalSupply?: SerializedBigNumber;
  tokenPriceVsQuote?: SerializedBigNumber;
  poolWeight?: SerializedBigNumber;
  userData?: SerializedFarmUserData;
}
// ...
export interface SerializedFarmsState {
  data: SerializedFarm[];
  loadArchivedFarmsData: boolean;
  userDataLoaded: boolean;
  loadingKeys: Record<string, boolean>;
  poolLength?: number;
  regularCakePerBlock?: number;
}